add visdrone functionality to main scripts
The Dataset class has code to properly edit the visdrone annotation
format to the expected format used in coco. Models.py was using indexing
with int arrays and this was changed to bools, to avoid annoying
messages about deprecation. train and test got changes to enable
visdrone dataset functionality but also to deal with errors in the
validation pipeline. It can now deal with situations where the models in
training is unable to predict any boxes above the threshold.