add code to test bounding box drawing code
Added a new script testAnnotations.py which loads visdrone images and
labels and prints the bounding box ground truths. This was to enable
the testing of the printbbboxes function in utils.py. printbbboxes.py
was created from code in the detect.py script to enable printing
bounding boxes from multiple sources (for validation/testing).