introduce cropping augmentation and improve logs
Train.py and datasets.py were modified to enable random cropping of
the images. This required removal of labels outside of the crop as well
as rescaling and renormalizing of labels that do fit. Logging was
changed to be in separate sections for loss and accuracy as well as
grouping each metric together. A single image that didnt fit the 500x500
crop size was removed from the image file list.